+++ tests/specs/ins/harvesting.e2e.js
@@ -1,4 +1,31 @@
+const now = require('../../lib/now.js')
+
 describe('ins harvesting', () => {
+  it('should submit PageActions', async () => {
+    const testUrl = await browser.testHandle.assetURL('instrumented.html')
+    await browser.url(testUrl)
+      .then(() => browser.waitForAgentLoad())
+
+    const [{ request: { body: { ins: pageActionsHarvest }, query } }] = await Promise.all([
+      browser.testHandle.expectIns(),
+      browser.execute(function () {
+        newrelic.addPageAction('DummyEvent', { free: 'tacos' })
+      })
+    ])
+
+    expect(pageActionsHarvest.length).toEqual(1)
+    let event = pageActionsHarvest[0]
+    expect(event.actionName).toEqual('DummyEvent')
+    expect(event.free).toEqual('tacos')
+    let receiptTime = now()
+    let relativeHarvestTime = query.rst
+    let estimatedPageLoad = receiptTime - relativeHarvestTime
+    let eventTimeSinceLoad = event.timeSinceLoad * 1000
+    let estimatedEventTime = eventTimeSinceLoad + estimatedPageLoad
+    expect(relativeHarvestTime > eventTimeSinceLoad).toEqual(true) //, 'harvest time (' + relativeHarvestTime + ') should always be bigger than event time (' + eventTimeSinceLoad + ')')
+    expect(estimatedEventTime < receiptTime).toEqual(true) //, 'estimated event time (' + estimatedEventTime + ') < receipt time (' + receiptTime + ')')
+  })
+
   it('NEWRELIC-9370: should not throw an exception when calling addPageAction with window.location before navigating', async () => {
     const testUrl = await browser.testHandle.assetURL('api/addPageAction-unload.html')
     await browser.url(testUrl)
